requirements-dev.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-extensions==2.2.9
djangorestframework==3.11.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
